+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableShortArray.kt
@@ -226,6 +226,11 @@ public value class ImmutableShortArray @PublishedApi internal constructor(
     public inline fun randomOrNull(): Short? = values.randomOrNull()
 
     /**
+     * See [ShortArray.randomOrNull]
+     */
+    public inline fun randomOrNull(random: Random): Short? = values.randomOrNull(random)
+
+    /**
      * See [ShortArray.iterator]
      */
     public inline operator fun iterator(): Iterator<Short> = values.iterator()